+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 2.12.0
 * added autolinks for posts and pages;
 * the view will automatically choose which layout is to be used;
+* splitted some frontend (default) css files;
 * from now, "frontend" is "default" and "backend" is "admin".
 
 ## 2.11 branch